% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_add_exp_des_col.R
\name{add_exp_des_col}
\alias{add_exp_des_col}
\title{add_exp_des_col}
\usage{
add_exp_des_col(
  data,
  d_exp_des,
  data_unit = "unit",
  d_exp_unit = "new_unit",
  col_add = c("rowNum", "colNum")
)
}
\arguments{
\item{data}{experimental data with experimental unit defined by the
argument "data_unit".}

\item{d_exp_des}{experimental design information (e.g. row, column)
corresponding to experimental unit defined by the argument
"d_exp_unit".}

\item{data_unit}{character string specifying the name of the column
of "data" that contain the identifier of the experimental unit.
Default = "unit".}

\item{d_exp_unit}{character string specifying the name of the column
of "d_exp_des" that contain the identifier of the experimental unit.
Default = "new_unit".}

\item{col_add}{column names of the column from "d_exp_des" that need
to be added to data. Default = c("rowNum", "colNum").}
}
\value{
Extended data.frame with experimental design information
}
\description{
Add extra columns with experimental design information
to a data.frame.
}
