% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Boxplot_TR.R
\name{Boxplot_TR}
\alias{Boxplot_TR}
\title{Boxplot_TR}
\usage{
Boxplot_TR(results, trait = "TRmax", main = "TR feature Boxplot")
}
\arguments{
\item{results}{object of class TRres obtained with \code{\link{TR_data_proc}}}

\item{trait}{Character string: one of "TRmax", "TR_AUC_total", or "TR_AUC_10h15h"}

\item{main}{Title of the graph.}
}
\value{
Boxplot
}
\description{
Boxplot of the average trait value over the time series per genotype.
}
\examples{

\dontrun{

# Transpiration rate data processing obtained with
# TR_data_proc
data("TR_res")

p <- Boxplot_TR(results = TR_res, trait = "TRmax")
p

}

}
