% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_check_timestamp_format.R
\name{check_timestamp_format}
\alias{check_timestamp_format}
\title{check_timestamp_format}
\usage{
check_timestamp_format(x)
}
\arguments{
\item{x}{character vector of timestamp (time point) values}
}
\value{
TRUE/FALSE value given the result of the text
}
\description{
Check if all elements from a vector of timestamp
match the %Y-%m-%d %H:%M:%S format.
}
