% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_Histogram_TR.R
\name{Histogram_TR}
\alias{Histogram_TR}
\title{Histogram_TR}
\usage{
Histogram_TR(
  results,
  trait = "TRmax",
  n_bins = 10,
  main = "TR feature Histogram"
)
}
\arguments{
\item{results}{object of class TRres obtained with \code{\link{TR_data_proc}}}

\item{trait}{Character string: one of "TRmax", "TR_AUC_total", or "TR_AUC_10h15h"}

\item{n_bins}{Number of bins. Default = 10.}

\item{main}{Title of the graph.}
}
\value{
Histogram plot
}
\description{
Histogram of the average trait value over the time series per genotype.
}
\examples{

\dontrun{

# Transpiration rate data processing obtained with
# TR_data_proc
data("TR_res")

p <- Histogram_TR(results = TR_res, trait = "TRmax")
p

}

}
