% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_LAI_coverage.R
\name{plot_LAI_coverage}
\alias{plot_LAI_coverage}
\title{plot_LAI_coverage}
\usage{
plot_LAI_coverage(results)
}
\arguments{
\item{results}{object of class TRres obtained with \code{\link{TR_data_proc}}}
}
\value{
matrix plot
}
\description{
A function to plot the coverage of LAI plant eye data
obtained with \code{\link{TR_data_proc}}.
}
\examples{

\dontrun{

# Transpiration rate data processing obtained with
# TR_data_proc
data("TR_res")

p <- plot_LAI_coverage(results = TR_res)
p

}

}
