% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_TR_time_series.R
\name{plot_TR_time_series}
\alias{plot_TR_time_series}
\title{plot_TR_time_series}
\usage{
plot_TR_time_series(
  results,
  trait = "TRmax",
  n_sector = NULL,
  color = FALSE,
  main = "TR feature TS plot"
)
}
\arguments{
\item{results}{object of class TRres obtained with \code{\link{TR_data_proc}}}

\item{trait}{Character string: one of "TRmax", "TR_AUC_total", or "TR_AUC_10h15h"}

\item{n_sector}{By default all genotypes are plotted, otherwise a sample of n_sector is randomly selected to be ploted.}

\item{color}{If color = TRUE, all genotypes get a different colour.}

\item{main}{Title of the graph.}
}
\value{
time series plot
}
\description{
A function to plot the daily time series of selected trait from
transpiration rate (TR) profile obtained with \code{\link{TR_data_proc}}.
}
\examples{

\dontrun{

# Transpiration rate data processing obtained with
# TR_data_proc
data("TR_res")

p <- plot_TR_time_series(results = TR_res, trait = "TRmax",
                         n_sector = NULL, color = FALSE)
p

}

}
