\name{sensor_data}
\alias{sensor_data}
\docType{data}
\title{
Example sensor data
}
\description{
Example sensor data
}
\usage{data("sensor_data")}
\format{
  A data frame with n observations on the following 2 variables.
  \describe{
    \item{\code{sector}}{a character vector describing the sector position}
    \item{\code{sensor}}{a character vector providing the sensor information.}
  }
}

\examples{
data(sensor_data)
}
\keyword{datasets}