\name{TR_res}
\alias{TR_res}
\docType{data}
\title{
  Example TR time series processing
}
\description{
  Example transpiration rate (TR) time series processing using 
   \code{\link{TR_data_proc}}.
}
\usage{data("TR_res")}
\format{
  A list of results obtained with \code{\link{TR_data_proc}}. 
}

\examples{
  data(TR_res)
}
\keyword{datasets}