% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_TR_res_get_traits.R
\name{TR_res_get_traits}
\alias{TR_res_get_traits}
\title{TR_res_get_traits}
\usage{
TR_res_get_traits(TR_res)
}
\arguments{
\item{TR_res}{object of class TRres obtained with \code{\link{TR_data_proc}}}
}
\value{
A \code{data.frame} with the following traits extracted for each sectors:

 \item{max_TR_av}{Average maximum daily transpiration over the selected period}
 \item{tot_AUC_sum}{Cumulated area under the transpiration curve over the selected period
 (Total amount transpired water)}
 \item{tot_AUC_av}{Average daily transpiration over the selected period}
 \item{AUC_10h15h_sum}{Cumulated area under the transpiration curve between 10h00 and 15h00
 over the selected period (Total amount transpired water between 10h00 and 15h00)}
 \item{AUC_10h15h_av}{Average daily transpiration between 10h00 and 15h00 over the selected period}
 \item{slope_07_maxTR_av}{Average daily transpiration ~ time slope between 07h00 and the time of max transpiration.}
 \item{slope_TR_VPD}{Linear regression coefficient Transpiration ~ VPD}
}
\description{
Function to extract traits from a TRres object
}
\examples{

data("TR_res")

TR_traits <- TR_res_get_traits(TR_res = TR_res)

}
