% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_TR_res_subset.R
\name{TR_res_subset}
\alias{TR_res_subset}
\title{TR_res_subset}
\usage{
TR_res_subset(TR_res, start_day, end_day)
}
\arguments{
\item{TR_res}{object of class TRres obtained with \code{\link{TR_data_proc}}}

\item{start_day}{\code{numeric} values specifying the first day of the subseted
TR_res.}

\item{end_day}{\code{numeric} values specifying the last day of the subseted
TR_res.}
}
\value{
subseted TRres object
}
\description{
A function to subset a TR_res object.
}
\examples{

data("TR_res")

TR_res <- TR_res_subset(TR_res = TR_res, start_day = 6, end_day = 21)

}
