% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_TR_res_to_pre_TP.R
\name{TR_res_to_pre_TP}
\alias{TR_res_to_pre_TP}
\title{TR_res_to_pre_TP}
\usage{
TR_res_to_pre_TP(TR_res)
}
\arguments{
\item{TR_res}{object of class TRres obtained with \code{\link{TR_data_proc}}}
}
\value{
The return value, if any, from executing the function.
}
\description{
Function to transform the results from a TR_res object
into a pre TP object format just before the addition of the
experimental design information.
}
