% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_TR_VPD_reg.R
\name{TR_VPD_reg}
\alias{TR_VPD_reg}
\title{TR_VPD_reg}
\usage{
TR_VPD_reg(TR_res, do_plot = FALSE, plot_loc, plot_name = NULL)
}
\arguments{
\item{TR_res}{object of class TRres obtained with \code{\link{TR_data_proc}}}

\item{do_plot}{\code{Logical} value specifying if the plot of the regression
should be saved. Default = FALSE.}

\item{plot_loc}{\code{Character} string specifying the path where the plot
should be saved.}

\item{plot_name}{Optional \code{character} specifying the name of the plot.
Default: TR_VPD_regression}
}
\value{
A \code{data.frame} with the Tr ~ VPD linear regression coefficients.
}
\description{
Calculate linear regression coefficient TR-VPD
}
\examples{

data("TR_res")

TR_VPD_reg_coeff <- TR_VPD_reg(TR_res = TR_res, do_plot = FALSE)

}
