% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_wth_data_proc.R
\name{wth_data_proc}
\alias{wth_data_proc}
\title{wth_data_proc}
\usage{
wth_data_proc(wth_data, sensor_data, skew_test = FALSE)
}
\arguments{
\item{wth_data}{\code{data.frame} representing the weather data information
containing the following columns (variables):

\describe{
 \item{\code{sensor}}{a character vector describing the sensor information}
 \item{\code{variable}}{a character vector indicating the type of weather information. Must be one of: Temperature (°C), Relative humidity (\%)), Windspeed average (m/s), Windspeed max (m/s), Solar radiation (W/(s*m²)), Precipitation (mm), or Wind direction (°).}
 \item{\code{timestamp}}{a character vector indicating the time of measurement in format yyyy-mm-dd hh:mm:ss.}
 \item{\code{value}}{a numeric vector indicating the variable value.}
}

The column names must be strictly equivalent to the one specified:
'sensor', 'variable', 'timestamp', 'value'. Default = NULL.}

\item{sensor_data}{\code{data.frame} representing the sensor data information
containing the following columns (variables):

\describe{
 \item{\code{sector}}{a character vector describing the sector position.}
 \item{\code{sensor}}{a character vector indicating the sensor information.}
}}

\item{skew_test}{\code{Logical value} specifying if skewness test
should be used to filter weather data. Default = FALSE.}
}
\value{
processed weather data
}
\description{
processing of the weather data
}
