% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TR_res_subset.R
\name{TR_res_subset}
\alias{TR_res_subset}
\title{Subsetting of transpiration rate processed data}
\usage{
TR_res_subset(TR_res, start_date = NULL, end_date = NULL)
}
\arguments{
\item{TR_res}{\code{List} of class \code{TRres} obtained with function \code{\link{TR_data_proc}}}

\item{start_date}{Default = NULL.}

\item{end_date}{Default = NULL.}
}
\value{
Return:

TR_res with only the results for the selected time range.
}
\description{
Subsetting of transpiration rate processed data
}
\examples{

Later


}
\author{
ICRISAT-GEMS
}
